prepare ZoomableImage component
